+++ src/MedicalAppointments/Patient.java
@@ -1,50 +1,13 @@
 package MedicalAppointments;
 
-public class Patient {
-    private String name;
-    private String email;
-    private String address;
-    private String phoneNombre;
+public class Patient extends User {
     private String birthday;
     private Double weight;
     private Double height;
     private String blood;
 
     Patient(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhoneNombre() {
-        return phoneNombre;
-    }
-
-    public void setPhoneNombre(String phoneNombre) {
-        this.phoneNombre = phoneNombre;
+        super(name, email);
     }
 
     public String getBirthday() {
